// src/app/shared/models/edilinkm.ts
import { GlobalVariables } from '../../core/models/globalvariables';

export class EdiLinkm {
    pkid: string;
    sourcetable: string;
    sourcetype: string;
    searchstring: string;
    searchvalue1: string;
    searchvalue2: string;
    sourceid: string;  
    sourcecode: string;  
    sourcename: string;  
    targetid: string;  
    targetdesc:string;
    
    branchcode: string;  
    rec_branch_code:string;
    
    rowdisplayed: boolean;
    rec_mode: string;
  
    _globalvariables: GlobalVariables;
  
}